Introduced basic HTML exporting.
There's a new Export submenu in Session, then the user selects Plain
Text or HTML. The output HTML is not ideal and totally relies on
QTextDocument::toHtml().

Improvement: dump QSyntaxHighlighter and perform highlighting manually
with HTML tags in Q(Plain)TextEdit. That way, the highlighting wil be
preserved, ass well as the empty break lines of the result display.

Further improvement: once Qt 4.4 becomes obsolete and we require Qt 4.5,
QTextDocumentWriter can be used to export to plain text, HTML, and ODF.