Add degree-minute-second conversion functions: dms, fromdms
"dms" takes an argument in degrees and formats it in degree-minute-second
format e.g. as 34* 24' 59.13" (with a degree symbol in place of the '*').
"fromdms" takes three arguments, the degrees, minutes, and seconds components
of an angle, and converts it to an angle in either degrees or radians,
depending on angle mode.

Possible issues:

- "dms" assumes its argument is in degrees, even in radian mode

- "fromdms" is not a very beautiful function name, and it doesn't seem to fit
    with the rest