Digit grouping
http://code.google.com/p/speedcrunch/issues/detail?id=58

Add support for digit grouping for all expressions and results.

A space between groups is added. That space is not a character, but rather a
font property. Therefore it is not copied when users copy and paste expressions.
Digits in the fractional and the exponent parts are also grouped.

The size of the groups depends on the radix:
    - 3 digits for decimal and octal
    - 4 digits for binary and hexadecimal

This feature shall be configurable by the user via Settings in a future commit.