Polish the merged copy behavior of editor and display
Now, text selection can exist on either editor or display,
never both. And the curently selected text (on either one
of the widgets) gets copied to clipboard, when the copy
key sequence or Ctrl-C is triggered.
In short both widgets act like one combined widget when it
comes to copying text to clipboard.

This path removed the automatic copy-on-selection feature
that the display had.