Add support for digits groups separators from user input (issue 434). Let the user choose which radix characters to use for parsing. Fix issue 151 when the number is parsed (rather than when it is converted).
This commit allow the token scanner to ignore digits groups separators in numbers from user input. In strict mode, the scanner only detects known separators (check http://en.wikipedia.org/wiki/Decimal_mark#Examples_of_use), while in non-strict mode, any character that is not alphanumeric nor a known operator is detected as a group separator. The latter permits to strip currency symbols or spaces in numbers from user input for instance.

Only 2 radix characters are recognized by the scanner (while more exist), but since it is not uncommon that some languages use one of these 2 characters as a radix character and the other as a digits groups separator, this commit adds an option so that the user can choose whether she wants the character not configured as a radix character to be detected as a radix character as well or as a group separator.