Remove support for hiding the menu
Maintaining this feature for Qt 4 and 5 on all platforms has become complex.
Also, removing the menu bar does not result in a better user experience.
From now on, let the window system decide if the bar is displayed in the
global menu, otherwise show it always as in normal applications.